move onSubmit to base component add template only component and rename login-form to login-template
