+++ ui/app/components/auth/page.hbs
@@ -55,7 +55,7 @@
           @onError={{fn (mut this.mfaErrors)}}
         />
       {{else}}
-        <Auth::LoginForm
+        <Auth::LoginTemplate
           @wrappedToken={{@wrappedToken}}
           @cluster={{@cluster}}
           @handleNamespaceUpdate={{this.handleNamespaceUpdate}}
